re-arrange file structure
