--- book.rb
@@ -1,18 +1,0 @@
-require './item'
-require 'date'
-
-class Book < Item
-  attr_accessor :publisher, :cover_state
-
-  def initialize(publisher, cover_state, date: Date.today)
-    super(date)
-    @publisher = publisher
-    @cover_state = cover_state
-  end
-
-  private
-
-  def can_be_archived?
-    super || @cover_state == 'bad'
-  end
-end
