+++ src/classes/game.rb
