+++ src/classes/music_album.rb
