+++ src/storage/labels.json
@@ -1,0 +1,20 @@
+[
+  {
+    "json_class": "Label",
+    "id": 559,
+    "title": "sasa",
+    "color": "da"
+  },
+  {
+    "json_class": "Label",
+    "id": 572,
+    "title": "dsdsd",
+    "color": "asd"
+  },
+  {
+    "json_class": "Label",
+    "id": 706,
+    "title": "dasd",
+    "color": "asd"
+  }
+]
